pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jupiter-llm"
version = "0.1.0"
description = "Framework de entrenamiento distribuido para modelos de lenguaje expertos en dominios específicos"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Jupiter Contributors"}
]
keywords = ["llm", "distributed-training", "mlx", "apple-silicon", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "tqdm>=4.65",
    "numpy>=1.24",
    "requests>=2.28",
    "aiohttp>=3.8",
    "aiofiles>=23.0",
    "beautifulsoup4>=4.12",
    "httpx>=0.24",
    "pydantic>=2.0",
    "huggingface-hub>=0.20",
    "tiktoken>=0.5",
    "sentencepiece>=0.1.99",
    "datasets>=2.14",
    "safetensors>=0.4",
    "mpi4py>=3.1",
]

[project.optional-dependencies]
mac = [
    "mlx>=0.20.0",
    "mlx-lm>=0.18.0",
]
nvidia = [
    "torch>=2.1",
    "transformers>=4.36",
    "accelerate>=0.25",
    "bitsandbytes>=0.41",
    "flash-attn>=2.3",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
all = [
    "jupiter-llm[mac,nvidia,dev]",
]

[project.scripts]
jupiter = "jupiter.cli:main"

[project.urls]
Homepage = "https://github.com/raym33/jupiter"
Repository = "https://github.com/raym33/jupiter"
Issues = "https://github.com/raym33/jupiter/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["jupiter*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
